requirements.txt
crewai
langchain_google_genai
python-dotenv
langchain
